Add explanation about date storage and range
This is the page that comes up when you google "mongodb date". The information about possible date range is now also on http://docs.mongodb.org/manual/reference/bson-types/#date, but that page is a lot harder to find. Alternatively, a link from this page to that page could be added.

Signed-off-by: kay <kay.kim@10gen.com>