changing findAndModify to update
See DOCS-3308 - changing findAndModify to update to hopefully reduce confusion out in the world that you need findAndModify to make an atomic update.   Obviously findAndModify could be added to this as a note that it is also atomic, but it returns the document that has been changed (which doesn't seem necessary in this case, since we have the _id for it).

Signed-off-by: kay <kay.kim@10gen.com>